+++ Code/Patches/LandValueOverhaulPatches.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="PropertyRenterSystemPatches.cs" company="algernon (K. Algernon A. Sheppard)">
+﻿// <copyright file="LandValueOverhaulPatches.cs" company="algernon (K. Algernon A. Sheppard)">
 // Copyright (c) algernon (K. Algernon A. Sheppard). All rights reserved.
 // Licensed under the Apache Licence, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // See LICENSE.txt file in the project root for full license information.
@@ -10,14 +10,12 @@ namespace PlopTheGrowables
     using System.Collections.Generic;
     using System.Reflection;
     using System.Reflection.Emit;
-    using Game.Simulation;
     using HarmonyLib;
 
     /// <summary>
-    /// Harmony patches for <see cref="PropertyRenterSystem"/> to implement building level locking.
+    /// Harmony patches for the <c>Land Value Overhaul</c> mod to implement building level locking.
     /// </summary>
-    [HarmonyPatch]
-    internal class PropertyRenterSystemPatches
+    internal class LandValueOverhaulPatches
     {
         /// <summary>
         /// Harmony transpiler for <c>PropertyRenterSystem.OnUpdate</c> to override game levelling.
@@ -25,8 +23,6 @@ namespace PlopTheGrowables
         /// <param name="instructions">Original ILCode.</param>
         /// <param name="original">Method being patched.</param>
         /// <returns>Modified ILCode.</returns>
-        [HarmonyPatch(typeof(PropertyRenterSystem), "OnUpdate")]
-        [HarmonyTranspiler]
         internal static IEnumerable<CodeInstruction> OnUpdateTranspiler(IEnumerable<CodeInstruction> instructions, MethodBase original)
         {
             Patcher.Instance.Log.Info($"Transpiling {original.DeclaringType}:{original.Name}");
@@ -34,8 +30,8 @@ namespace PlopTheGrowables
             // Levelup and Leveldown job types and local indices.
             int levelUpJobIndex = int.MaxValue;
             int levelDownJobIndex = int.MaxValue;
-            Type levelUpJobType = Type.GetType("Game.Simulation.PropertyRenterSystem+LevelupJob,Game", true);
-            Type levelDownJobType = Type.GetType("Game.Simulation.PropertyRenterSystem+LeveldownJob,Game", true);
+            Type levelUpJobType = Type.GetType("LandValueOverhaul.Systems.PropertyRenterSystem+LevelupJob,LandValueOverhaul", true);
+            Type levelDownJobType = Type.GetType("LandValueOverhaul.Systems.PropertyRenterSystem+LeveldownJob,LandValueOverhaul", true);
 
             // Get local indices.
             foreach (LocalVariableInfo localVarInfo in original.GetMethodBody().LocalVariables)
@@ -64,14 +60,14 @@ namespace PlopTheGrowables
                     {
                         Mod.Instance.Log.Debug($"Skipping local {localBuilder.LocalIndex} from {instruction.opcode} {instruction.operand}");
 
-                        // Skip forward until stloc.3, indicating the end of the job creation block.
-                        while (instruction.opcode != OpCodes.Stloc_3)
+                        // Skip forward until stloc.s 4, indicating the end of the job creation block.
+                        while (!(instruction.opcode == OpCodes.Stloc_S && instruction.operand is LocalBuilder localBuilder2 && localBuilder2.LocalIndex == 4))
                         {
                             instructionEnumerator.MoveNext();
                             instruction = instructionEnumerator.Current;
                         }
 
-                        // Skip current instruction (the stloc.3);
+                        // Skip current instruction (the stloc.s 4);
                         Mod.Instance.Log.Debug($"resuming after {instruction.opcode} {instruction.operand}");
                         continue;
                     }